fix: isolate Cloud terminal close compatibility
